+++ src/ui.rs
@@ -88,12 +88,12 @@ impl UI {
     }
 }
 
-pub fn get_hits_from_nth_loop(user_hits: &Vec<UserHit>, desired_loop_idx: i32) -> Vec<UserHit> {
+pub fn get_hits_from_nth_loop(user_hits: &Vec<UserHit>, desired_loop_idx: usize) -> Vec<UserHit> {
     let last_loop_hits: Vec<UserHit> = user_hits
         .iter()
         .filter(|hit| {
             // include hits from just before start of loop (back to 0 - MISS), since those could be early or on-time hits
-            let loop_num_for_hit = ((hit.clock_tick + MISS_MARGIN) / BEATS_PER_LOOP) as i32;
+            let loop_num_for_hit = ((hit.clock_tick + MISS_MARGIN) / BEATS_PER_LOOP) as usize;
             loop_num_for_hit == desired_loop_idx
         })
         .map(|hit| hit.clone())
